update  the doc of cube class refactor the cube class position
