+++ cp2kdata/__init__.py
@@ -1,4 +1,4 @@
 from .output import Cp2kOutput
 from .utils import get_opt_cell_stc
 from .pdos import Cp2kPdos
-from .cube import Cp2kCube
+from .cube.cube import Cp2kCube
