+++ docs/cube/MAV_compare.png
