+++ docs/cube/PAV_compare.png
